Address Review
As per a recent review
on this branch, the zelbet
constants have been moved
into plookup folder and
exposed by the module
