+++ src/plookup/table/hash_tables/constants.rs
@@ -11,6 +11,7 @@
 // Currently making the s_i usize.
 
 use bigint::U256 as u256;
+use dusk_bls12_381::BlsScalar;
 
 /// Constant V
 pub const V: usize = 643;
@@ -685,4 +686,202 @@ pub const SBOX_U256: [u256; 659] = [
     u256([547, 0, 0, 0]),
     u256([100, 0, 0, 0]),
     u256([346, 0, 0, 0]),
+];
+
+/// decomposition = [s_n, s_{n-1} ..., s_1]
+pub const DECOMPOSITION_S_I: [BlsScalar; 27] = [
+    BlsScalar([693, 0, 0, 0]),
+    BlsScalar([696, 0, 0, 0]),
+    BlsScalar([694, 0, 0, 0]),
+    BlsScalar([668, 0, 0, 0]),
+    BlsScalar([679, 0, 0, 0]),
+    BlsScalar([695, 0, 0, 0]),
+    BlsScalar([691, 0, 0, 0]),
+    BlsScalar([693, 0, 0, 0]),
+    BlsScalar([700, 0, 0, 0]),
+    BlsScalar([688, 0, 0, 0]),
+    BlsScalar([700, 0, 0, 0]),
+    BlsScalar([694, 0, 0, 0]),
+    BlsScalar([701, 0, 0, 0]),
+    BlsScalar([694, 0, 0, 0]),
+    BlsScalar([699, 0, 0, 0]),
+    BlsScalar([701, 0, 0, 0]),
+    BlsScalar([701, 0, 0, 0]),
+    BlsScalar([701, 0, 0, 0]),
+    BlsScalar([695, 0, 0, 0]),
+    BlsScalar([698, 0, 0, 0]),
+    BlsScalar([697, 0, 0, 0]),
+    BlsScalar([703, 0, 0, 0]),
+    BlsScalar([702, 0, 0, 0]),
+    BlsScalar([691, 0, 0, 0]),
+    BlsScalar([688, 0, 0, 0]),
+    BlsScalar([703, 0, 0, 0]),
+    BlsScalar([679, 0, 0, 0]),
+];
+
+/// decomposition_inverses (are in Montgomery form) = [s_n^{-1}, ...,
+/// s_1^{-1}]
+pub const INVERSES_S_I: [BlsScalar; 27] = [
+    BlsScalar([
+        10221572469640980478,
+        9996602938199176322,
+        12629687422955549043,
+        8006457294428823986,
+    ]),
+    BlsScalar([
+        2544378491691269391,
+        4278332345953639965,
+        11874303680749211767,
+        2426939821791091123,
+    ]),
+    BlsScalar([
+        744249036390430597,
+        12703574479313007857,
+        12941063472551889101,
+        7898626469976166322,
+    ]),
+    BlsScalar([
+        12813307260754568983,
+        13401737678607148669,
+        16259656273855951252,
+        77635945376597827,
+    ]),
+    BlsScalar([
+        5868183680725469139,
+        13963385237577580741,
+        2771541177262918091,
+        2057109021922450213,
+    ]),
+    BlsScalar([
+        12076645397819930386,
+        3513102646780106621,
+        5511421539288173570,
+        5771838565320455168,
+    ]),
+    BlsScalar([
+        6887496337358022873,
+        10595820730184048433,
+        5009316404333812521,
+        5382134171948568183,
+    ]),
+    BlsScalar([
+        10221572469640980478,
+        9996602938199176322,
+        12629687422955549043,
+        8006457294428823986,
+    ]),
+    BlsScalar([
+        4216398644020393879,
+        8184599923631624209,
+        8370160424198379966,
+        2747212268616577148,
+    ]),
+    BlsScalar([
+        9008875010444565957,
+        5372543878254038396,
+        4977331236853060682,
+        4494972308657887955,
+    ]),
+    BlsScalar([
+        4216398644020393879,
+        8184599923631624209,
+        8370160424198379966,
+        2747212268616577148,
+    ]),
+    BlsScalar([
+        744249036390430597,
+        12703574479313007857,
+        12941063472551889101,
+        7898626469976166322,
+    ]),
+    BlsScalar([
+        14604768842820473715,
+        14849596265064785286,
+        7748405401081812358,
+        5829685384640365743,
+    ]),
+    BlsScalar([
+        744249036390430597,
+        12703574479313007857,
+        12941063472551889101,
+        7898626469976166322,
+    ]),
+    BlsScalar([
+        158341149692136564,
+        11915096076612138202,
+        992774181608407826,
+        2273115756299035808,
+    ]),
+    BlsScalar([
+        14604768842820473715,
+        14849596265064785286,
+        7748405401081812358,
+        5829685384640365743,
+    ]),
+    BlsScalar([
+        14604768842820473715,
+        14849596265064785286,
+        7748405401081812358,
+        5829685384640365743,
+    ]),
+    BlsScalar([
+        14604768842820473715,
+        14849596265064785286,
+        7748405401081812358,
+        5829685384640365743,
+    ]),
+    BlsScalar([
+        12076645397819930386,
+        3513102646780106621,
+        5511421539288173570,
+        5771838565320455168,
+    ]),
+    BlsScalar([
+        14693968055083782227,
+        11284659261154016671,
+        1375836735433574310,
+        7087423626300479469,
+    ]),
+    BlsScalar([
+        8098570567349679003,
+        8435273170816927395,
+        2562595664368193229,
+        3130570460078912387,
+    ]),
+    BlsScalar([
+        9026571779888472938,
+        4361001158901360622,
+        8564805700575463153,
+        287658769533324946,
+    ]),
+    BlsScalar([
+        15293454485951598931,
+        15506724579931865072,
+        2628833535283587331,
+        4905112046230959809,
+    ]),
+    BlsScalar([
+        6887496337358022873,
+        10595820730184048433,
+        5009316404333812521,
+        5382134171948568183,
+    ]),
+    BlsScalar([
+        9008875010444565957,
+        5372543878254038396,
+        4977331236853060682,
+        4494972308657887955,
+    ]),
+    BlsScalar([
+        9026571779888472938,
+        4361001158901360622,
+        8564805700575463153,
+        287658769533324946,
+    ]),
+    BlsScalar([
+        5868183680725469139,
+        13963385237577580741,
+        2771541177262918091,
+        2057109021922450213,
+    ]),
 ];
