+++ src/plookup/table/hash_tables/mod.rs
@@ -12,4 +12,4 @@ pub mod constants;
 /// Preliminary file building reinforced concrete's tables
 pub mod tables;
 
-pub use constants::{N, S, T_S, V};
+pub use constants::{DECOMPOSITION_S_I, INVERSES_S_I, N, S, SBOX_U256, T_S, V};
